requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain-community==0.3.8
json-repair==0.30.2
pydantic==2.10.2
openai==1.55.3